+++ tests/tests-method-initialize.R
@@ -3,7 +3,8 @@ library(data.cube)
 
 # level ----
 
-lvl = level$new(x = data.table(a = rep(1:6,2), b = letters[1:3], d = letters[1:2], z = NA), key = "a", properties = c("b","d"))
+dt = data.table(a = rep(1:6,2), b = letters[1:3], d = letters[1:2], z = NA)
+lvl = level$new(x = dt, key = "a", properties = c("b","d"))
 stopifnot(
     inherits(lvl, "level"),
     identical(lvl$key, "a"),
@@ -13,15 +14,19 @@ stopifnot(
 
 # hierarchy ----
 
-hrh1 = hierarchy$new(x = data.table(a = rep(1:6,2), b = letters[1:3], d = letters[1:2], z = NA), key = "a", levels = list("b" = character(), "a" = c("b")))
-hrh2 = hierarchy$new(x = data.table(a = rep(1:6,2), b = letters[1:3], d = letters[1:2], z = NA), key = "a", levels = list("d" = character(), "a" = c("d")))
-stopifnot(TRUE) # TO DO
-# sequence processing of levels in hierarchy from top to bottom keeping parent references
+hrh1 = hierarchy$new(levels = list("b" = character(), "a" = c("b")))
+hrh2 = hierarchy$new(levels = list("d" = character(), "a" = c("d")))
+stopifnot(
+    is.list(hrh1$levels),
+    length(hrh1$levels) == 2L,
+    identical(unname(sapply(hrh1$levels, length)), 0:1)
+)
 
 # dimension ----
 
-dim = dimension$new(
-    x = data.table(a = rep(1:6,2), b = letters[1:3], d = letters[1:2], z = NA),
+dt = data.table(a = rep(1:6,2), b = letters[1:3], d = letters[1:2], z = NA)
+ddim = dimension$new(
+    x = dt,
     key = "a",
     hierarchies = list(
         list("b" = character(), "a" = c("b")),
@@ -29,50 +34,41 @@ dim = dimension$new(
     )
 )
 stopifnot(
-    identical(base::dim(dim$data), c(6L, 3L)),
-    identical(dim$key, "a"),
-    length(dim$hierarchies) == 2L,
-    sapply(dim$hierarchies, function(x) length(x$levels)) == 2L,
-    c(sapply(dim$hierarchies, function(x) names(x$levels))) == c("b","a","d","a")
+    identical(dim(ddim$data), c(6L, 3L)),
+    identical(ddim$key, "a"),
+    length(ddim$hierarchies) == 2L,
+    sapply(ddim$hierarchies, function(x) length(x$levels)) == 2L,
+    c(sapply(ddim$hierarchies, function(x) names(x$levels))) == c("b","a","d","a")
 )
 
 # time dimension
-X = populate_star(N = 1e5, surrogate.keys = FALSE)
+X = populate_star(N = 1e5, surrogate.keys = FALSE, hierarchies = TRUE)
 time = dimension$new(X$dims$time,
                      key = "time_date",
-                     hierarchies = list(
-                         "monthly" = list(
-                             "time_year" = character(),
-                             "time_quarter" = c("time_quarter_name"),
-                             "time_month" = c("time_month_name"),
-                             "time_date" = c("time_month","time_quarter","time_year")
-                         ),
-                         "weekly" = list(
-                             "time_year" = character(),
-                             "time_week" = character(),
-                             "time_date" = c("time_week","time_year")
-                         )
-                     ))
+                     hierarchies = X$hierarchies$time)
 stopifnot(
     inherits(time, "dimension"),
     names(time$hierarchies) == c("monthly","weekly"),
     time$hierarchies$monthly$key == "time_date",
     # all leafs in levels are normalized data.tables
-    unlist(lapply(time$hierarchies, function(hrh) lapply(hrh$levels, function(lvl) is.data.table(lvl$data))))
+    unlist(lapply(time$levels, function(lvl) is.data.table(lvl$data)))
 )
-sprintf("size of all tables in 'time' dimension: %.3f MB.", sum(unlist(lapply(time$hierarchies, function(hrh) lapply(hrh$levels, function(lvl) as.numeric(object.size(lvl$data))))))/(1024L^2L))
 rm(X)
 
 # measure ----
 
 m = measure$new(x = "z", label = "score etc.")
 m = measure$new(x = "z", label = "score etc.", fun.aggregate = "sum", na.rm = TRUE)
-m$format()
-stopifnot(TRUE) # TO DO
+
+stopifnot(
+    is.measure(m),
+    is.language(m$format())
+)
 
 # fact ----
 
-ff = fact$new(x = data.table(a = rep(1:6,2), b = letters[1:3], d = letters[1:2], z = 1:12*sin(1:12)),
+dt = data.table(a = rep(1:6,2), b = letters[1:3], d = letters[1:2], z = 1:12*sin(1:12))
+ff = fact$new(x = dt,
               id.vars = c("a","b","d"),
               measure.vars = "z")
 stopifnot(
@@ -81,44 +77,25 @@ stopifnot(
 
 # data.cube ----
 
-X = populate_star(N = 1e5, surrogate.keys = FALSE)
+X = populate_star(N = 1e5, surrogate.keys = FALSE, hierarchies = TRUE)
 time = dimension$new(X$dims$time,
                      key = "time_date",
-                     hierarchies = list(
-                         "monthly" = list(
-                             "time_year" = character(),
-                             "time_quarter" = c("time_quarter_name"),
-                             "time_month" = c("time_month_name"),
-                             "time_date" = c("time_month","time_quarter","time_year")
-                         ),
-                         "weekly" = list(
-                             "time_year" = character(),
-                             "time_week" = character(),
-                             "time_date" = c("time_week","time_year")
-                         )
-                     ))
+                     hierarchies = X$hierarchies$time)
 geog = dimension$new(X$dims$geography,
                      key = "geog_abb",
-                     hierarchies = list(
-                         list(
-                             "geog_region_name" = character(),
-                             "geog_division_name" = character(),
-                             "geog_abb" = c("geog_name","geog_division_name","geog_region_name")
-                         )
-                     ))
+                     hierarchies = X$hierarchies$geography)
 ff = fact$new(x = X$fact$sales,
               id.vars = c("geog_abb","time_date"),
               measure.vars = c("amount","value"),
               fun.aggregate = "sum",
               na.rm = TRUE)
-
 dc = data.cube$new(fact = ff, dimensions = list(time = time, geography = geog))
 stopifnot(
     is.data.table(dc$fact$data),
     # Normalization
-    identical(dim(dc$dimensions$geography$hierarchies[[1L]]$levels$geog_abb$data), c(50L, 4L)),
-    identical(dim(dc$dimensions$geography$hierarchies[[1L]]$levels$geog_region_name$data), c(4L, 1L)),
-    identical(dim(dc$dimensions$time$hierarchies[[1L]]$levels$time_month$data), c(12L, 2L)),
-    identical(dim(dc$dimensions$time$hierarchies[[1L]]$levels$time_year$data), c(5L, 1L)),
-    identical(dim(dc$dimensions$time$hierarchies[[1L]]$levels$time_date$data), c(1826L, 4L))
+    identical(dim(dc$dimensions$geography$levels$geog_abb$data), c(50L, 4L)),
+    identical(dim(dc$dimensions$geography$levels$geog_region_name$data), c(4L, 1L)),
+    identical(dim(dc$dimensions$time$levels$time_month$data), c(12L, 2L)),
+    identical(dim(dc$dimensions$time$levels$time_year$data), c(5L, 1L)),
+    identical(dim(dc$dimensions$time$levels$time_date$data), c(1826L, 5L))
 )